alien não treme mais ao chegar ao destino; FALTA ADD MINION

+++ include/Action.h
@@ -7,7 +7,7 @@ class Action {
         enum ActionType {MOVE=0, SHOOT};
         ActionType type;
         Vec2 pos;
-        Action(ActionType, double, double);
+        Action(ActionType, int, int);
 };
